Fix for "generated" files in top directory
Latest changes had only been done to these files, not to their "sources"
(which happen to be verbatim copies).

The "generated" files are now ignored by Git and the changes have been
applied to the "sources".

TODO: decrease confusion further by simplifying the build process and
doing away with the "generated" files.